ImageData - Make thread variables a property.
Second round of refactor for thread variables. First, make the variable
a property on the image_data parent class to ensure all inheriting
classes have this defined. The _init_ call will set that to None and the
first access will copy the defined BASE_THREAD_VARIABLES in there. The
latter constant changed to a set as well to ensure immutability once
defined on the class.